extracted imagespec struct
