+++ lua/kubectl/client/rust/dao/daemonset.rs
@@ -3,35 +3,10 @@ use k8s_openapi::{
     serde_json::{json, Map, Value},
 };
 use kube::api::{Api, Patch, PatchParams};
-use mlua::{FromLua, Lua, Result as LuaResult, Value as LuaValue};
+use mlua::{Lua, Result as LuaResult};
 use std::collections::HashMap;
 
-use crate::with_client;
-
-#[derive(Debug, Clone)]
-pub struct ImageSpec {
-    pub name: String,
-    pub image: String,
-    pub init: bool,
-}
-
-impl FromLua for ImageSpec {
-    fn from_lua(value: LuaValue, _lua: &Lua) -> LuaResult<Self> {
-        match value {
-            LuaValue::Table(t) => {
-                let name: String = t.get("name")?;
-                let image: String = t.get("image")?;
-                let init: bool = t.get("init")?;
-                Ok(ImageSpec { name, image, init })
-            }
-            _ => Err(mlua::Error::FromLuaConversionError {
-                from: value.type_name(),
-                to: "ImageSpec".to_string(),
-                message: None,
-            }),
-        }
-    }
-}
+use crate::{structs::ImageSpec, with_client};
 
 pub fn set_images(_lua: &Lua, args: (String, String, Vec<ImageSpec>)) -> LuaResult<String> {
     let (daemonset_name, namespace, images) = args;
