+++ lua/kubectl/client/types.lua
@@ -7,4 +7,11 @@
 --- @field portforward_start fun(kind: string, name: string, namespace: string, local_port: number, remote_port: number)
 --- @field portforward_list fun()
 --- @field portforward_stop fun(id: number)
+--- @field daemonset_set_images fun(name: string, ns: string, image_spec: {} )
 --- @field deployment_set_images fun(name: string, ns: string, image_spec: {} )
+--- @field statefulset_set_images fun(name: string, ns: string, image_spec: {} )
+--- @field create_job_from_cronjob fun(j_name: string, ns: string, cj_name: string, dry_run: boolean )
+--- @field suspend_cronjob fun(cj_name: string, ns: string, suspend: boolean)
+--- @field uncordon_node fun(name: string)
+--- @field cordon_node fun(name: string)
+--- @field get_config fun()
